Added parallelization.py module to provide functions for parallelizing computations using multiple CPU cores.
The parallelize_computation function allows users to apply a given function to a list of data items concurrently, leveraging multiprocessing to achieve parallel execution. This module can be used to speed up computations that can be processed independently on multiple CPU cores. Users can import and use the parallelize_computation function in other parts of the codebase to parallelize computations efficiently.